+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - 18.09.2018.
 
-* TODO: Describe initial release.
+initial version with these API services:
+- add a new product
+- add a new product image
+- get an existing product